% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMissingValues.R
\name{fillMissingValues}
\alias{fillMissingValues}
\title{Fill Missing Values}
\usage{
fillMissingValues(Data)
}
\arguments{
\item{Data}{The historical data of the roads of Thessaloniki}
}
\value{
A data frame with all the historical data without missing values
}
\description{
This function fills the missing values from the data.
}
\details{
This function returns a data frame without missing values.
}
\examples{
SpecLink <- loadDataSpecLink("163204843","1", X163204843_1)
x <- fillMissingValues(SpecLink)

}
\seealso{
\code{\link{loadData}}
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis, Charalampos Bratsas
}
