% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load traffic data}
\usage{
loadData()
}
\value{
Returns a data frame.
}
\description{
This function loads the traffic data.
}
\details{
This function returns a data frame with the traffic data of the roads of Thessaloniki ordered by the roads.
}
\seealso{

}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
