% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataSpecLink.R
\name{loadDataSpecLink}
\alias{loadDataSpecLink}
\title{Load data for a specific road of Thessaloniki}
\usage{
loadDataSpecLink(Link_id, direction, Data)
}
\arguments{
\item{Link_id}{A character with the id of the road needed}

\item{direction}{The direction of the road}

\item{Data}{The historical data of the roads of Thessaloniki}
}
\value{
A data frame with the data of a specific road
}
\description{
This function extracts the data of one road of Thessaloniki.
}
\details{
This function returns a data frame with the historical data of a specific road.
}
\seealso{
\code{\link{loadData}}
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
