% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCR.R
\name{PredictionCR}
\alias{PredictionCR}
\title{Prediction}
\usage{
PredictionCR(List,NNOut,predict)
}
\arguments{
\item{List}{A list with the following components: trainset, testset, MinMaxFromScaling}

\item{NNOut}{The train model}

\item{predict}{The value to be predicted}
}
\value{
The predicted average speed of the road
}
\description{
This function predicts the average speed of the road.
}
\details{
This function returns the predicted average speed.
}
\seealso{
\code{\link{PreProcessingLink}}, \code{\link{TrainCR}}
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
