% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessingLink.R
\name{PreProcessingLink}
\alias{PreProcessingLink}
\title{PreProcessing second model}
\usage{
PreProcessingLink(DataList)
}
\arguments{
\item{DataList}{A list with the following components: trainData, testData, 
trainDataWide, cormat}
}
\value{
A list with the following components:

\itemize{
\item trainset The trainset for the model
\item testset The testset to be predict
\item Minimum The min values of each column of the initial dataset
\item Maximum The max values of each column of the initial dataset
}
}
\description{
This function processes the data.
}
\details{
This function returns as a list object the parameters needed to train the model and predict.
}
\examples{
SpecLink <- loadDataSpecLink("163204843","1", X163204843_1)
x <- fillMissingValues(SpecLink)
datetime <- "2017-01-27 14:00:00" 
newData <- fillMissingDates (x, datetime)
DataList <- loadTrainTest (newData, datetime, "Mean_speed")
List <- PreProcessingLink(DataList)

}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
