% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/163204843_1.R
\name{X163204843_1}
\alias{X163204843_1}
\title{Sample data from Traffic BDE}
\format{RData file}
\source{
TrafficBDE
}
\description{
Sample data of the traffic data of the road with Lik id "163204843" and direction = "1"
\itemize{
  \item The Link id of the road
  \item The direction of the road
  \item The date and time of the recorded arguments
  \item The min speed each time
  \item The max speed each time
  \item The mean speed each time
  \item The standard deviation of the speed
  \item The skewness of the speed
  \item The kurtosis of the speed
  \item The entries each time
  \item The unique entries each time
  }
}
