% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMissingDates.R
\name{fillMissingDates}
\alias{fillMissingDates}
\title{Fill Missing Dates}
\usage{
fillMissingDates(Data, datetime)
}
\arguments{
\item{Data}{The historical data}

\item{datetime}{The datetime wanted}
}
\value{
A data frame with all the historical data between the first date and the date wanted.
}
\description{
This function fills the missing dates from the data.
}
\details{
This function returns a data frame without missing dates.
}
\examples{
SpecLink <- loadDataSpecLink("163204843","1", X163204843_1)
x <- fillMissingValues(SpecLink)
datetime <- "2017-01-27 14:00:00" 
newData <- fillMissingDates (x, datetime)

}
\seealso{
\code{\link{loadData}}, \code{\link{fillMissingValues}}
}
\author{
Aikaterini Chatzopoulou, Charalampos Bratsas
}
