% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainCR.R
\name{TrainCR}
\alias{TrainCR}
\title{Train}
\usage{
TrainCR(List,predict)
}
\arguments{
\item{List}{A list with the following components: trainset, testset, Min, Max}

\item{predict}{The value to be predicted}
}
\value{
The train model
}
\description{
This function trains the model.
}
\details{
This function returns the trained model.
}
\examples{
\dontrun{
SpecLink <- loadDataSpecLink("163204843","1", X163204843_1)
x <- fillMissingValues(SpecLink)
datetime <- "2017-01-27 14:00:00" 
newData <- fillMissingDates (x, datetime)
DataList <- loadTrainTest (newData, datetime, "Mean_speed")
List <- PreProcessingLink(DataList)
NNout <- TrainCR (List,"Mean_speed")}

}
\seealso{
\code{\link{PreProcessingLink}}
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
